requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
confluent-kafka
